+++ archetype/spring-data-sample/src/main/resources/archetype-resources/src/test/java/sample/SampleTest.java
@@ -31,7 +31,7 @@ import org.springframework.test.context.*;
 import org.springframework.context.annotation.*;
 import org.springframework.beans.factory.annotation.*;
 import org.springframework.data.jpa.repository.config.EnableJpaRepositories;
-import com.blazebit.persistence.view.impl.spring.EnableEntityViews;
+import com.blazebit.persistence.view.spring.impl.EnableEntityViews;
 import com.blazebit.persistence.spring.data.impl.repository.EntityViewRepositoryFactoryBean;
 
 @RunWith(SpringJUnit4ClassRunner.class)