+++ examples/spring-showcase/src/test/java/com/blazebit/persistence/examples/spring/SpringShowcaseTest.java
@@ -17,27 +17,19 @@
 package com.blazebit.persistence.examples.spring;
 
 import com.blazebit.persistence.examples.spi.Showcase;
-import com.blazebit.persistence.view.impl.spring.EnableEntityViews;
 import org.junit.After;
 import org.junit.Before;
-import org.junit.BeforeClass;
 import org.junit.Test;
 import org.junit.runner.RunWith;
 import org.junit.runners.Parameterized;
 import org.springframework.beans.factory.config.AutowireCapableBeanFactory;
-import org.springframework.context.annotation.ComponentScan;
-import org.springframework.context.annotation.Configuration;
-import org.springframework.context.annotation.ImportResource;
 import org.springframework.test.annotation.DirtiesContext;
 import org.springframework.test.context.ContextConfiguration;
 import org.springframework.test.context.TestContextManager;
-import org.springframework.test.context.junit4.SpringJUnit4ClassRunner;
 import org.springframework.test.context.support.DependencyInjectionTestExecutionListener;
-import org.springframework.test.context.support.DirtiesContextBeforeModesTestExecutionListener;
 import org.springframework.test.context.support.DirtiesContextTestExecutionListener;
 
 import javax.inject.Inject;
-import java.util.Iterator;
 import java.util.ServiceLoader;
 
 /**